commands: move config code to config.rs
